+++ src/main/java/com/zerobase/babdeusilbun/config/SchedulerConfig.java
@@ -1,4 +1,4 @@
-package com.zerobase.babdeusilbun.scheduler;
+package com.zerobase.babdeusilbun.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
