+++ src/main/java/com/zerobase/babdeusilbun/security/service/impl/JwtValidationServiceImpl.java
@@ -4,7 +4,7 @@ import static com.zerobase.babdeusilbun.exception.ErrorCode.JWT_TOKEN_INVALID;
 
 import com.zerobase.babdeusilbun.exception.CustomException;
 import com.zerobase.babdeusilbun.security.service.JwtValidationService;
-import com.zerobase.babdeusilbun.security.util.JwtComponent;
+import com.zerobase.babdeusilbun.security.component.JwtComponent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
