+++ src/main/java/com/zerobase/babdeusilbun/security/service/impl/SignServiceImpl.java
@@ -10,7 +10,7 @@ import static com.zerobase.babdeusilbun.exception.ErrorCode.USER_MEETING_STILL_L
 import static com.zerobase.babdeusilbun.exception.ErrorCode.USER_NOT_FOUND;
 import static com.zerobase.babdeusilbun.exception.ErrorCode.USER_POINT_NOT_EMPTY;
 import static com.zerobase.babdeusilbun.exception.ErrorCode.USER_WITHDRAWAL;
-import static com.zerobase.babdeusilbun.security.constants.SecurityConstantsUtil.*;
+import static com.zerobase.babdeusilbun.security.util.SecurityConstantsUtil.*;
 import static com.zerobase.babdeusilbun.security.redis.RedisKeyUtil.jwtBlackListKey;
 import static com.zerobase.babdeusilbun.security.redis.RedisKeyUtil.refreshTokenKey;
 import static com.zerobase.babdeusilbun.security.type.Role.ROLE_ENTREPRENEUR;
@@ -32,7 +32,6 @@ import com.zerobase.babdeusilbun.repository.MeetingRepository;
 import com.zerobase.babdeusilbun.repository.PurchaseRepository;
 import com.zerobase.babdeusilbun.repository.SchoolRepository;
 import com.zerobase.babdeusilbun.repository.UserRepository;
-import com.zerobase.babdeusilbun.security.constants.SecurityConstantsUtil;
 import com.zerobase.babdeusilbun.security.dto.SignRequest;
 import com.zerobase.babdeusilbun.security.dto.SignRequest.BusinessSignUp;
 import com.zerobase.babdeusilbun.security.dto.SignRequest.SignIn;
@@ -41,9 +40,7 @@ import com.zerobase.babdeusilbun.security.dto.WithdrawalRequest;
 import com.zerobase.babdeusilbun.exception.CustomException;
 import com.zerobase.babdeusilbun.security.service.SignService;
 import com.zerobase.babdeusilbun.security.type.Role;
-import com.zerobase.babdeusilbun.security.util.JwtComponent;
-import com.zerobase.babdeusilbun.util.NicknameUtil;
-import jakarta.servlet.http.HttpServletResponse;
+import com.zerobase.babdeusilbun.security.component.JwtComponent;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
